+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 **Changes in version 1.10**
 
+ - Changed MphUtils.getAllGroups() to return a map of ID and MphGroup.
+ - MphGroup names changed from "Year Name" to "Name (Year)".
+ - Updated Hematopoietic2010 files with latest data from SEER API Client.
+ - Added missing space to message "Unable to apply Rule".
+ - Standardized and combined "Unable to apply Rule" warning messages.
+ - Fixed 2007 Malignant Brain Group Charts to contain multiple branches.
  - Updated Commons Lang library from version 3.4 to version 3.7.
  - Updated CSV library from version 2.4 to version 4.2.
-
+ 
 **Changes in version 1.9**
 
  - Added a generic rule to return same primary when inputs have same and valid properties.